requirements.txt
Django==4.0.5
django_filter==21.1
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0
environ==1.0
